ai-service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
tensorflow==2.18.1
opencv-python==4.10.0.84
pillow==11.0.0
numpy==1.26.4
scikit-image==0.22.0
imagehash==4.3.1
exifread==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
redis==5.0.1
requests==2.31.0
